requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
ffmpeg==1.4
gTTS==2.4.0
idna==3.4
pydub==0.25.1
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.0
urllib3==2.0.7
